Updated database functions and added test files
Modified some of the existing database functions and added functions for Login/Logout, Recent Trips, and Suggested Purchases. Also added test files to show examples of function usage. "tylerDBTests.php" shows possible outcomes of Recent Trips and Suggested Purchases, and also has *commented* code that will test the Login function and redirect to the "tylerLoginTest.php" page if successful. Otherwise, the "tylerLoginTest.php" page will show 'Not logged in' if manually navigated to (and not previously logged in).